Added test_list_tenant non-admin test to v2/test_tenants.py
test_list_tenant test uses non-admin user's token and public
(non-admin) Identity API v2 client to make a request to list
tenants as a user action. Test checks that user can see only
that tenants that he was bound to and cannot log in to tenant
that is not in his list.

Change-Id: Iabf3a474e70f87bc494d87d6333d668d52de2968